merged: backported warning fixes from linuxport
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@12435 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/UnrarXLib/model.cpp
@@ -402,7 +402,7 @@ inline bool PPM_CONTEXT::decodeSymbol1(ModelPPM *Model)
   STATE* p=U.Stats;
   int i, HiCnt;
   int count=Model->Coder.GetCurrentCount();
-  if (count>=Model->Coder.SubRange.scale)
+  if ((uint)count>=Model->Coder.SubRange.scale)
     return(false);
   if (count < (HiCnt=p->Freq)) 
   {
@@ -491,7 +491,7 @@ inline bool PPM_CONTEXT::decodeSymbol2(ModelPPM *Model)
   } while ( --i );
   Model->Coder.SubRange.scale += HiCnt;
   count=Model->Coder.GetCurrentCount();
-  if (count>=Model->Coder.SubRange.scale)
+  if ((uint)count>=Model->Coder.SubRange.scale)
     return(false);
   p=*(pps=ps);
   if (count < HiCnt)
